+++ src/components/Editor.vue
@@ -1,0 +1,48 @@
+<template>
+  <div>
+    <vue-editor
+      :editorToolbar="toolbarContent"
+      v-model="content"
+    />
+  </div>
+</template>
+
+<script>
+import { VueEditor } from 'vue2-editor'
+
+export default {
+  components: {
+    VueEditor
+  },
+  // name: 'ComponentName',
+  props: [ 'text' ],
+  data () {
+    return {
+      content: this.text,
+      toolbarContent: [
+        [{ 'header': [] }],
+        ['bold', 'italic', 'underline', 'strike'],
+        ['blockquote'],
+        [{ 'list': 'ordered' }, { 'list': 'bullet' }],
+        ['clean']
+      ]
+    }
+  }
+}
+</script>
+
+<style>
+
+.ql-picker-label {
+  color: #bbb !important;
+}
+
+.ql-fill {
+  fill: #bbb !important;
+}
+
+.ql-stroke {
+  stroke: #bbb !important;
+}
+
+</style>