+++ src/layouts/builder.vue
@@ -2,7 +2,7 @@
   <q-layout> <!-- Be sure to play with the Layout demo on docs -->
 
     <!-- (Optional) The Header -->
-    <q-layout-header>
+    <q-layout-header reveal>
       <q-toolbar>
         <q-btn
           flat
@@ -12,14 +12,13 @@
           @click="leftDrawer = !leftDrawer"
         />
         <q-toolbar-title>
-          Header
-          <span slot="subtitle">Subtile</span>
+          <img src="statics/logo.png" style="max-height: 40px" />
         </q-toolbar-title>
       </q-toolbar>
     </q-layout-header>
 
     <!-- (Optional) The Footer -->
-    <q-layout-footer>
+    <!-- <q-layout-footer>
       <q-toolbar>
         <q-btn
           flat
@@ -33,7 +32,7 @@
           <span slot="subtitle">Subtile</span>
         </q-toolbar-title>
       </q-toolbar>
-    </q-layout-footer>
+    </q-layout-footer> -->
 
     <!-- (Optional) A Drawer; you can add one more with side="right" or change this one's side -->
     <q-layout-drawer
@@ -43,10 +42,49 @@
     >
       <!-- QScrollArea is optional -->
       <q-scroll-area class="fit q-pa-sm">
+        <q-item to="/dashboard">
+          <q-item-side icon="fa-home" />
+          <q-item-main label="Dashboard" />
+        </q-item>
+        <q-item-separator />
+        <q-item>
+          <q-item-main label="REAL Curriculum" />
+        </q-item>
+        <q-item :to="{ name: 'rlist', params: { type: 'series' }}">
+          <q-item-side icon="fa-folder" />
+          <q-item-main label="Series" />
+        </q-item>
+        <q-item-separator />
+        <q-item>
+          <q-item-main label="Other Curriculum" />
+        </q-item>
+        <q-item :to="{ name: 'olist', params: { type: 'series' }}">
+          <q-item-side icon="fa-folder" />
+          <q-item-main label="Series" />
+        </q-item>
+        <q-item :to="{ name: 'olist', params: { type: 'lessons' }}">
+          <q-item-side icon="fa-file" />
+          <q-item-main label="Lessons" />
+        </q-item>
+        <q-item-separator />
+        <q-item>
+          <q-item-side icon="fa-cog" />
+          <q-item-main label="Settings" />
+        </q-item>
         <q-item link @click.native="logout">
-          <q-item-side icon="fa-sign-out fab" />
+          <q-item-side icon="fa-sign-out-alt fab" />
           <q-item-main label="Log Out" />
         </q-item>
+      </q-scroll-area>
+    </q-layout-drawer>
+
+    <q-layout-drawer
+      side="right"
+      v-model="rightDrawer"
+      content-class="bg-grey-9"
+    >
+      <q-scroll-area>
+        <resources />
       </q-scroll-area>
     </q-layout-drawer>
 
@@ -59,11 +97,17 @@
 </template>
 
 <script>
+import Resources from 'components/Resources.vue'
+
 export default {
+  components: {
+    Resources
+  },
   // name: 'LayoutName',
   data () {
     return {
-      leftDrawer: true
+      leftDrawer: true,
+      rightDrawer: true
     }
   },
   methods: {