+++ CHANGELOG.md
@@ -1,7 +1,20 @@
 # Changelog
 All notable changes for this project will be documented in this file.
 
-## 0.0.2 - 2018-03-19
+## 0.0.3 - 2018-03-21
+### Added
+- database plugin
+- add functionality to database
+- can add Lessons and Series to other
+- OList for listing other series and lessons
+- RList for listing REAL series (possibly lessons)
+- played with Edtiors a bit
+- adding BibleRefs to an olesson will present the passages
+
+### Issues
+- No idea how to get resources from NQ to here...
+
+## 0.0.2 - 2018-03-20
 ### Added
 - Login page
 - Dashboard page